Fix move image bug
Before when moving the image the canvas would reset. This was because
the `useDraw` hook was still in an old closure with an old artboard
state.

Trying to add `setPosition` from `usePosition` as a dependency did
not work, because then the event listener for the mouse would not
trigger properly.

In the end with the current model, it's not needed to keep setPosition
inside react hooks, so I just moved it out and now it works.

The model for the dataflow is like this:

interface -> firebase -> react -> interfact